initial frontend newProject modal refactor
