+++ frontend/src/components/v2/projects/index.tsx
@@ -1,0 +1,1 @@
+export { NewProjectModal } from "./NewProjectModal";
