+++ frontend/src/layouts/AppLayout/AppLayout.tsx
@@ -6,7 +6,6 @@
 /* eslint-disable func-names */
 
 import { useEffect, useMemo, useState } from "react";
-import { Controller, useForm } from "react-hook-form";
 import { useTranslation } from "react-i18next";
 import Link from "next/link";
 import { useRouter } from "next/router";
@@ -21,66 +20,48 @@ import {
   faEnvelope,
   faInfinity,
   faInfo,
-  faInfoCircle,
   faMobile,
   faPlus,
   faQuestion,
   faStar as faSolidStar
 } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { yupResolver } from "@hookform/resolvers/yup";
 import { DropdownMenuTrigger } from "@radix-ui/react-dropdown-menu";
 import { twMerge } from "tailwind-merge";
-import * as yup from "yup";
 
 import { createNotification } from "@app/components/notifications";
 import { OrgPermissionCan } from "@app/components/permissions";
 import { tempLocalStorage } from "@app/components/utilities/checks/tempLocalStorage";
 import SecurityClient from "@app/components/utilities/SecurityClient";
 import {
-  Accordion,
-  AccordionContent,
-  AccordionItem,
-  AccordionTrigger,
   Button,
-  Checkbox,
   DropdownMenu,
   DropdownMenuContent,
   DropdownMenuItem,
-  FormControl,
-  Input,
   Menu,
   MenuItem,
-  Modal,
-  ModalContent,
   Select,
   SelectItem,
   UpgradePlanModal
 } from "@app/components/v2";
+import { NewProjectModal } from "@app/components/v2/projects/NewProjectModal";
 import {
   OrgPermissionActions,
   OrgPermissionSubjects,
   useOrganization,
-  useOrgPermission,
   useSubscription,
   useUser,
   useWorkspace
 } from "@app/context";
 import { usePopUp, useToggle } from "@app/hooks";
 import {
-  fetchOrgUsers,
-  useAddUserToWsNonE2EE,
-  useCreateWorkspace,
   useGetAccessRequestsCount,
-  useGetExternalKmsList,
   useGetOrgTrialUrl,
   useGetSecretApprovalRequestCount,
   useLogoutUser,
   useSelectOrganization
 } from "@app/hooks/api";
 import { MfaMethod } from "@app/hooks/api/auth/types";
-import { INTERNAL_KMS_KEY_ID } from "@app/hooks/api/kms/types";
-import { InfisicalProjectTemplate, useListProjectTemplates } from "@app/hooks/api/projectTemplates";
 import { Workspace } from "@app/hooks/api/types";
 import { useUpdateUserProjectFavorites } from "@app/hooks/api/users/mutation";
 import { useGetUserProjectFavorites } from "@app/hooks/api/users/queries";
@@ -119,20 +100,6 @@ const supportOptions = [
   ]
 ];
 
-const formSchema = yup.object({
-  name: yup
-    .string()
-    .required()
-    .label("Project Name")
-    .trim()
-    .max(64, "Too long, maximum length is 64 characters"),
-  addMembers: yup.bool().required().label("Add Members"),
-  kmsKeyId: yup.string().label("KMS Key ID"),
-  template: yup.string().label("Project Template Name")
-});
-
-type TAddProjectFormData = yup.InferType<typeof formSchema>;
-
 export const AppLayout = ({ children }: LayoutProps) => {
   const router = useRouter();
 
@@ -165,10 +132,6 @@ export const AppLayout = ({ children }: LayoutProps) => {
 
   const { data: secretApprovalReqCount } = useGetSecretApprovalRequestCount({ workspaceId });
   const { data: accessApprovalRequestCount } = useGetAccessRequestsCount({ projectSlug });
-  const { permission } = useOrgPermission();
-  const { data: externalKmsList } = useGetExternalKmsList(currentOrg?.id!, {
-    enabled: permission.can(OrgPermissionActions.Read, OrgPermissionSubjects.Kms)
-  });
 
   const pendingRequestsCount = useMemo(() => {
     return (secretApprovalReqCount?.open || 0) + (accessApprovalRequestCount?.pendingCount || 0);
@@ -178,27 +141,24 @@ export const AppLayout = ({ children }: LayoutProps) => {
     ? subscription.workspacesUsed < subscription.workspaceLimit
     : true;
 
-  const createWs = useCreateWorkspace();
-  const addUsersToProject = useAddUserToWsNonE2EE();
-
   const infisicalPlatformVersion = process.env.NEXT_PUBLIC_INFISICAL_PLATFORM_VERSION;
 
-  const { popUp, handlePopUpOpen, handlePopUpClose, handlePopUpToggle } = usePopUp([
+  const { popUp, handlePopUpOpen, handlePopUpToggle } = usePopUp([
     "addNewWs",
     "upgradePlan",
     "createOrg"
   ] as const);
-  const {
-    control,
-    formState: { isSubmitting },
-    reset,
-    handleSubmit
-  } = useForm<TAddProjectFormData>({
-    resolver: yupResolver(formSchema),
-    defaultValues: {
-      kmsKeyId: INTERNAL_KMS_KEY_ID
-    }
-  });
+  // const {
+  //   control,
+  //   formState: { isSubmitting },
+  //   reset,
+  //   handleSubmit
+  // } = useForm<TAddProjectFormData>({
+  //   resolver: yupResolver(formSchema),
+  //   defaultValues: {
+  //     kmsKeyId: INTERNAL_KMS_KEY_ID
+  //   }
+  // });
 
   const { t } = useTranslation();
 
@@ -280,58 +240,6 @@ export const AppLayout = ({ children }: LayoutProps) => {
     };
     putUserInOrg();
   }, [router.query.id]);
-
-  const canReadProjectTemplates = permission.can(
-    OrgPermissionActions.Read,
-    OrgPermissionSubjects.ProjectTemplates
-  );
-
-  const { data: projectTemplates = [] } = useListProjectTemplates({
-    enabled: Boolean(canReadProjectTemplates && subscription?.projectTemplates)
-  });
-
-  const onCreateProject = async ({ name, addMembers, kmsKeyId, template }: TAddProjectFormData) => {
-    // type check
-    if (!currentOrg) return;
-    if (!user) return;
-    try {
-      const {
-        data: {
-          project: { id: newProjectId }
-        }
-      } = await createWs.mutateAsync({
-        projectName: name,
-        kmsKeyId: kmsKeyId !== INTERNAL_KMS_KEY_ID ? kmsKeyId : undefined,
-        template
-      });
-
-      if (addMembers) {
-        const orgUsers = await fetchOrgUsers(currentOrg.id);
-        await addUsersToProject.mutateAsync({
-          usernames: orgUsers
-            .filter(
-              (member) => member.user.username !== user.username && member.status === "accepted"
-            )
-            .map((member) => member.user.username),
-          projectId: newProjectId,
-          orgId: currentOrg.id
-        });
-      }
-
-      // eslint-disable-next-line no-promise-executor-return -- We do this because the function returns too fast, which sometimes causes an error when the user is redirected.
-      await new Promise((resolve) => setTimeout(resolve, 2_000));
-
-      // eslint-disable-next-line no-promise-executor-return -- We do this because the function returns too fast, which sometimes causes an error when the user is redirected.
-      await new Promise((resolve) => setTimeout(resolve, 2_000));
-
-      createNotification({ text: "Project created", type: "success" });
-      handlePopUpClose("addNewWs");
-      router.push(`/project/${newProjectId}/secrets/overview`);
-    } catch (err) {
-      console.error(err);
-      createNotification({ text: "Failed to create project", type: "error" });
-    }
-  };
 
   const addProjectToFavorites = async (projectId: string) => {
     try {
@@ -916,176 +824,10 @@ export const AppLayout = ({ children }: LayoutProps) => {
               </div>
             </nav>
           </aside>
-          <Modal
+          <NewProjectModal
             isOpen={popUp.addNewWs.isOpen}
-            onOpenChange={(isModalOpen) => {
-              handlePopUpToggle("addNewWs", isModalOpen);
-              reset();
-            }}
-          >
-            <ModalContent
-              title="Create a new project"
-              subTitle="This project will contain your secrets and configurations."
-            >
-              <form onSubmit={handleSubmit(onCreateProject)}>
-                <div className="flex gap-2">
-                  <Controller
-                    control={control}
-                    name="name"
-                    defaultValue=""
-                    render={({ field, fieldState: { error } }) => (
-                      <FormControl
-                        label="Project Name"
-                        isError={Boolean(error)}
-                        errorText={error?.message}
-                        className="flex-1"
-                      >
-                        <Input {...field} placeholder="Type your project name" />
-                      </FormControl>
-                    )}
-                  />
-                  <Controller
-                    control={control}
-                    name="template"
-                    render={({ field: { value, onChange } }) => (
-                      <OrgPermissionCan
-                        I={OrgPermissionActions.Read}
-                        a={OrgPermissionSubjects.ProjectTemplates}
-                      >
-                        {(isAllowed) => (
-                          <FormControl
-                            label="Project Template"
-                            icon={<FontAwesomeIcon icon={faInfoCircle} size="sm" />}
-                            tooltipText={
-                              <>
-                                <p>
-                                  Create this project from a template to provision it with custom
-                                  environments and roles.
-                                </p>
-                                {subscription && !subscription.projectTemplates && (
-                                  <p className="pt-2">Project templates are a paid feature.</p>
-                                )}
-                              </>
-                            }
-                          >
-                            <Select
-                              defaultValue={InfisicalProjectTemplate.Default}
-                              placeholder={InfisicalProjectTemplate.Default}
-                              isDisabled={!isAllowed || !subscription?.projectTemplates}
-                              value={value}
-                              onValueChange={onChange}
-                              className="w-44"
-                            >
-                              {projectTemplates.length
-                                ? projectTemplates.map((template) => (
-                                    <SelectItem key={template.id} value={template.name}>
-                                      {template.name}
-                                    </SelectItem>
-                                  ))
-                                : Object.values(InfisicalProjectTemplate).map((template) => (
-                                    <SelectItem key={template} value={template}>
-                                      {template}
-                                    </SelectItem>
-                                  ))}
-                            </Select>
-                          </FormControl>
-                        )}
-                      </OrgPermissionCan>
-                    )}
-                  />
-                </div>
-                <div className="mt-4 pl-1">
-                  <Controller
-                    control={control}
-                    name="addMembers"
-                    defaultValue={false}
-                    render={({ field: { onBlur, value, onChange } }) => (
-                      <OrgPermissionCan
-                        I={OrgPermissionActions.Read}
-                        a={OrgPermissionSubjects.Member}
-                      >
-                        {(isAllowed) => (
-                          <div>
-                            <Checkbox
-                              id="add-project-layout"
-                              isChecked={value}
-                              onCheckedChange={onChange}
-                              isDisabled={!isAllowed}
-                              onBlur={onBlur}
-                            >
-                              Add all members of my organization to this project
-                            </Checkbox>
-                          </div>
-                        )}
-                      </OrgPermissionCan>
-                    )}
-                  />
-                </div>
-                <div className="mt-14 flex">
-                  <Accordion type="single" collapsible className="w-full">
-                    <AccordionItem
-                      value="advance-settings"
-                      className="data-[state=open]:border-none"
-                    >
-                      <AccordionTrigger className="h-fit flex-none pl-1 text-sm">
-                        <div className="order-1 ml-3">Advanced Settings</div>
-                      </AccordionTrigger>
-                      <AccordionContent>
-                        <Controller
-                          render={({ field: { onChange, ...field }, fieldState: { error } }) => (
-                            <FormControl
-                              errorText={error?.message}
-                              isError={Boolean(error)}
-                              label="KMS"
-                            >
-                              <Select
-                                {...field}
-                                onValueChange={(e) => {
-                                  onChange(e);
-                                }}
-                                className="mb-12 w-full bg-mineshaft-600"
-                              >
-                                <SelectItem value={INTERNAL_KMS_KEY_ID} key="kms-internal">
-                                  Default Infisical KMS
-                                </SelectItem>
-                                {externalKmsList?.map((kms) => (
-                                  <SelectItem value={kms.id} key={`kms-${kms.id}`}>
-                                    {kms.name}
-                                  </SelectItem>
-                                ))}
-                              </Select>
-                            </FormControl>
-                          )}
-                          control={control}
-                          name="kmsKeyId"
-                        />
-                      </AccordionContent>
-                    </AccordionItem>
-                  </Accordion>
-                  <div className="absolute right-0 bottom-0 mr-6 mb-6 flex items-start justify-end">
-                    <Button
-                      key="layout-cancel-create-project"
-                      onClick={() => handlePopUpClose("addNewWs")}
-                      colorSchema="secondary"
-                      variant="plain"
-                      className="py-2"
-                    >
-                      Cancel
-                    </Button>
-                    <Button
-                      isDisabled={isSubmitting}
-                      isLoading={isSubmitting}
-                      key="layout-create-project-submit"
-                      className="ml-4"
-                      type="submit"
-                    >
-                      Create Project
-                    </Button>
-                  </div>
-                </div>
-              </form>
-            </ModalContent>
-          </Modal>
+            onClose={() => handlePopUpToggle("addNewWs", false)}
+          />
           <UpgradePlanModal
             isOpen={popUp.upgradePlan.isOpen}
             onOpenChange={(isOpen) => handlePopUpToggle("upgradePlan", isOpen)}
